Fix backward compatibility issue in signed JARs
The default in JDK 1.8.0_101 and later is to use SHA-256 as the
timestamp authority (TSA) digest algorithm, which makes the signed JARs
incompatible with earlier JRE releases.  Apparently that behavior was
also back-ported to OpenJDK 1.7.0_121, which this Docker image
previously used when building the 64-bit hybrid Java/native TurboVNC
Viewer.  The easiest solution is to build with JDK 1.8.0_92, which is
the last version that used SHA-1 as the default TSA digest algorithm.